move learningSection to pages/learning
